# specs/001-/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: 攝影作品展示網站 API
  description: 支援年份、作品集與照片管理的 Next.js API Routes
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API Routes
components:
  schemas:
    Year:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          example: "2024"
        order_index:
          type: string
          example: "2024.0"
        status:
          type: string
          enum: [draft, published]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, label, order_index, status]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
      required: [error, message]
  securitySchemes:
    CloudflareAccess:
      type: http
      scheme: bearer
      description: "Cloudflare Access JWT token"
paths:
...
  /years:
    get:
      summary: 獲取年份列表
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, all]
          example: published
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功獲取年份列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Year'

    post:
      summary: 創建新年份
      security:
        - CloudflareAccess: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  example: "2025"
                order_index:
                  type: string
                  example: "2025.0"
                status:
                  type: string
                  enum: [draft, published]
                  default: draft
              required: [label]
      responses:
        '201':
          description: 成功創建年份
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Year'
        '400':
          description: 請求資料無效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /years/{year_id}:
    get:
      summary: 獲取特定年份資訊
      parameters:
        - name: year_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功獲取年份資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Year'
        '404':
          description: 年份不存在

    put:
      summary: 更新年份資訊
      security:
        - CloudflareAccess: []
      parameters:
        - name: year_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                order_index:
                  type: string
                status:
                  type: string
                  enum: [draft, published]
      responses:
        '200':
          description: 成功更新年份
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Year'

    delete:
      summary: 刪除年份
      security:
        - CloudflareAccess: []
      parameters:
        - name: year_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: force
          in: query
          schema:
            type: boolean
            default: false
          description: "是否強制刪除含有作品集的年份"
      responses:
        '204':
          description: 成功刪除年份
        '409':
          description: 年份包含作品集，無法刪除

  # 作品集管理 API
...